+++ src/layer.cpp
@@ -9,13 +9,13 @@ Layer::Layer(size_t in, size_t out, std::mt19937 &rng)
   double std = std::sqrt(2.0 / in);
   std::normal_distribution<double> dist(0.0, std);
   for (size_t i = 0; i < W.noOfElements(); ++i)
-    W.flat(i) = dist(rng);
+    W.flat(i) = static_cast<float>(dist(rng));
   for (size_t i = 0; i < b.noOfElements(); ++i)
-    b.flat(i) = 0.0;
+    b.flat(i) = 0.0f;
   for (size_t i = 0; i < vW.noOfElements(); ++i)
-    vW.flat(i) = 0.0;
+    vW.flat(i) = 0.0f;
   for (size_t i = 0; i < vb.noOfElements(); ++i)
-    vb.flat(i) = 0.0;
+    vb.flat(i) = 0.0f;
 }
 
 Tensor Layer::forward(const Tensor &X) {
@@ -32,14 +32,14 @@ void Layer::step(double lr, double beta)
 {
     for (size_t i = 0; i < W.noOfElements(); ++i)
     {
-        vW.flat(i) = beta * vW.flat(i) + dW.flat(i);
-        W.flat(i) -= lr * vW.flat(i);
+        vW.flat(i) = static_cast<float>(beta * vW.flat(i) + dW.flat(i));
+        W.flat(i) -= static_cast<float>(lr * vW.flat(i));
     }
 
     for (size_t i = 0; i < b.noOfElements(); ++i)
     {
-        vb.flat(i) = beta * vb.flat(i) + db.flat(i);
-        b.flat(i) -= lr * vb.flat(i);
+        vb.flat(i) = static_cast<float>(beta * vb.flat(i) + db.flat(i));
+        b.flat(i) -= static_cast<float>(lr * vb.flat(i));
     }
 }
